Move `combinations` to new file for test utilities
