+++ src/xtbml/utils/__init__.py
@@ -4,4 +4,4 @@ Collection of utility functions.
 
 from .misc import is_int_list, is_str_list, symbol2number
 from .timing import Timers, timings
-from .torch import combinations, load_from_npz, real_atoms, real_pairs, t2int
+from .torch import load_from_npz, real_atoms, real_pairs, t2int
